+++ lib/config.dart
@@ -24,6 +24,7 @@ import 'package:yaml/yaml.dart';
 
 import '/util/log.dart';
 import '/util/platform_utils.dart';
+import 'domain/model/user.dart';
 import 'pubspec.g.dart';
 import 'routes.dart';
 import 'util/ios_utils.dart';
@@ -168,6 +169,12 @@ class Config {
   /// Indicator whether logs from `stdout` and `stderr` streams should be
   /// redirected.
   static bool redirectStdOut = true;
+
+  /// [UserId] of the [User]-support.
+  static String supportId = '7GgdGcrEEuLRWMDJakuDkI';
+
+  /// [UserId]s of the [User]s that should be considered as supports.
+  static List<String> supportIds = [];
 
   /// Initializes this [Config] by applying values from the following sources
   /// (in the following order):
@@ -335,6 +342,16 @@ class Config {
       Log.warning('init() -> announcements failed: $e', 'Config');
     }
 
+    supportId = const bool.hasEnvironment('SOCAPP_SUPPORT_ID')
+        ? const String.fromEnvironment('SOCAPP_SUPPORT_ID')
+        : (document['support']?['id'] ?? supportId);
+
+    final String? ids = const bool.hasEnvironment('SOCAPP_SUPPORT_IDS')
+        ? const String.fromEnvironment('SOCAPP_SUPPORT_IDS')
+        : document['support']?['ids'];
+
+    supportIds = ids == null ? [supportId, 'gapopa', 'tapopa'] : ids.split(',');
+
     // Change default values to browser's location on web platform.
     if (PlatformUtils.isWeb) {
       if (document['server']?['http']?['url'] == null &&
@@ -433,6 +450,12 @@ class Config {
               Log.warning('init() -> announcements failed: $e', 'Config');
             }
 
+            supportId = remote['support']?['id'] ?? supportId;
+            final String? ids = remote['support']?['ids'];
+            if (ids != null) {
+              supportIds = ids.split(',');
+            }
+
             origin = url;
           }
         }
@@ -473,6 +496,10 @@ class Config {
       });
     }
   }
+
+  /// Indicates whether the provided [UserId] is within [supportIds].
+  static bool isSupport(UserId id) =>
+      [Config.supportId, ...Config.supportIds].contains(id.val);
 }
 
 /// [title] with a [body] intended to be used as an announcement to make to the